requirements.txt
PyQt5>=5.14
pydicom
pillow